requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
supabase==2.15.3
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
python-multipart==0.0.6 